Ignore leading SECURITY_PROTOCOL:// in bootstrap_servers
